+++ integration-tests/tests/src/tests/index.ts
@@ -31,3 +31,4 @@ import "./dictionary";
 import "./credentials/anonymous";
 import "./sync/mixed";
 import "./sync/flexible";
+import "./sync/asymmetric";